Normaliza modalidade
