+++ datasets/tests/test_adapters.py
@@ -1,6 +1,13 @@
 from datetime import date, datetime
 
-from datasets.adapters import to_bid, to_contract, to_expense, to_revenue
+import pytest
+from datasets.adapters import (
+    normalize_category,
+    to_bid,
+    to_contract,
+    to_expense,
+    to_revenue,
+)
 
 
 def test_save_expense_from_csv():
@@ -147,7 +154,7 @@ def test_adapt_from_csv_data_to_revenue():
         "DTPUBLICACAO": "1/1/2014",
         "DTREGISTRO": "1/1/2014",
         "TIPOREC": "ORC",
-        "MODALIDADE": "RECEBIMENTO",
+        "MODALIDADE": "Repasse a Prefeitura indenização",
         "DSRECEITA": "TESTE DE RECEITA",
         "VALOR": "123131,00",
         "FONTE": "PREFEITURA",
@@ -162,7 +169,7 @@ def test_adapt_from_csv_data_to_revenue():
         "published_at": date(2014, 1, 1),
         "registered_at": date(2014, 1, 1),
         "revenue_type": "orcamentaria",
-        "modality": "recebimento",
+        "modality": "repasse_a_prefeitura_indenizacao",
         "description": "TESTE DE RECEITA",
         "value": 123131.00,
         "resource": "prefeitura",
@@ -185,3 +192,17 @@ def test_adapt_from_csv_data_to_revenue():
     assert revenue_obj.legal_status == expected_revenue["legal_status"]
     assert revenue_obj.destination == expected_revenue["destination"]
     assert revenue_obj.excluded == expected_revenue["excluded"]
+
+
+@pytest.mark.parametrize(
+    "string,expected_category",
+    [
+        ("Transferencia de Duodecimo", "transferencia_de_duodecimo"),
+        ("RECEBIMENTO", "recebimento"),
+        ("ORÇAMENTO", "orcamento"),
+        ("Repasse a Prefeitura indenização", "repasse_a_prefeitura_indenizacao"),
+        ("TRANSFERENCIA DUODECIMO", "transferencia_duodecimo"),
+    ],
+)
+def test_normalize_category(string, expected_category):
+    assert normalize_category(string) == expected_category
